feat: finishing the majority of the project

+++ src/controllers/event.controller.ts
@@ -4,20 +4,15 @@ import { EventDocument, EventInput } from "../models/event.model";
 import bcrypt from "bcrypt";
 import userService from "../services/user.service";
 import { UserDocument } from "../models/user.model";
+import { Console } from "console";
 
 class EventController {
     
     public async create(req: Request, res: Response){
         try {
             
-            const user: UserDocument | null = await userService.findById(req.params.id);
-
-            if(user?.role === "organizador") {    
-                const event: EventDocument = await EventService.create(req.params.id, req.body as EventInput);
-                return res.status(201).json(event);
-            } else {
-                return res.status(401).json({message: "Not authorized to create events"});
-            }
+            const event: EventDocument = await EventService.create(req.params.id, req.body as EventInput);
+            return res.status(201).json(event);
     
         } catch(error) {
             return res.status(500).json(error)
@@ -71,31 +66,46 @@ class EventController {
         }
     }
 
-    public async update(req:Request, res:Response){
+    public async update(req: Request, res: Response) {
         try {
-            const eventExists: EventDocument | null = await EventService.findByTitle(req.params.title);
-
-            if(!eventExists){
-                return res.status(404).json({message: "Event not found"});
+            const eventId = req.params.idEvent;
+            const updatedEventData = req.body;
+            const eventExists: EventDocument | null = await EventService.findById(eventId);
+            if (!eventExists) {
+                return res.status(404).json({ message: "Event doesn't exist" });
             }
-
-            if(req.body.password)
-                req.body.password = await bcrypt.hash(req.body.password, 10);
+            const user: UserDocument | null = await userService.findById(req.params.id);
+            if (user && eventExists.createdBy.toString() !== user._id.toString()) {
+                return res.status(403).json({ message: "Event is not owned by the user" });
+            }
+            const updatedEvent = await EventService.update(eventId, updatedEventData);
+    
+            return res.json(updatedEvent);
             
-            const updateUser = await EventService.update(req.params.title, req.body)
-            return res.status(200).json(updateUser);
         } catch (error) {
-            return res.status(500).json(error)
-        }
+            return res.status(500).json({ message: "Internal server error" });
+        }   
     }
 
-    public async delete(req:Request, res:Response){
+    public async delete(req: Request, res: Response) {
         try {
+            const eventExists: EventDocument | null = await EventService.findById(req.params.idEvent);
+            const user: UserDocument | null = await userService.findById(req.params.id);
             const idEvent = req.params.idEvent;
-            const events = await EventService.delete(idEvent);
-            res.json(events);
+            
+            if (!eventExists) {
+                return res.status(404).json({ message: "Event doesn't exist" });
+            }
+    
+            if (user && eventExists.createdBy.toString() === user._id.toString()) {
+                await EventService.delete(idEvent);
+                return res.json({ message: "Event deleted successfully" });
+            }
+    
+            return res.status(403).json({ message: "Event is not owned by the user" });
+            
         } catch (error) {
-            return res.status(500).json(error)
+            return res.status(500).json({ message: "Internal server error"});
         }   
     }
 
@@ -103,7 +113,16 @@ class EventController {
         res.status(201).json({"id": req.params.id});
     }
 
-
+    public async getAttendees(req: Request, res: Response) {
+        const eventId = req.params.eventId;
+    
+        try {
+            const attendees = await EventService.getAttendees(eventId);
+            res.json(attendees);
+        } catch (error) {
+            res.status(500).json({ message: 'Error retrieving the list of event attendees'});
+        }
+    }
 }
 
 export default new EventController();
